reorganize to limit viz dependencies
